requirements.txt
wxPython
Pillow
mutagen